+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/messages/factory/MessageContentFactory.kt
@@ -78,6 +78,7 @@ public open class MessageContentFactory {
             currentUser = currentUser,
             onLongItemClick = onLongItemClick,
             onLinkClick = onLinkClick,
+            onUserMentionClick = onUserMentionClick,
         )
     }